if path exists remove it

+++ src/media/services/import_dump/import_from_dump_service.py
@@ -19,8 +19,9 @@ class ImportFromDumpService:
         self.dump_to_database_service = DumpToDatabaseService()
 
     def import_from_dump(self, site: str, import_all: bool = False) -> int:
-        shutil.rmtree(self.EXTRACT_DIR)
         self._init(site)
+        if os.path.exists(self.EXTRACT_DIR):
+            shutil.rmtree(self.EXTRACT_DIR)
 
         # 1. Download zip file
         if self._should_download_zip():
